+++ src/sensors/KellerParent.h
@@ -40,11 +40,11 @@ class KellerParent : public Sensor
 public:
     KellerParent(byte modbusAddress, Stream* stream,
              int8_t powerPin, int8_t powerPin2, int8_t enablePin = -1, uint8_t measurementsToAverage = 1,
-             kellerModel model = OTHER, const char *sensName = "Keller-Sensor", int numVariables = 3,
+             kellerModel model = OTHER, const char *sensName = "Keller-Sensor", uint8_t numVariables = 3,
              uint32_t warmUpTime_ms = 500, uint32_t stabilizationTime_ms = 5000, uint32_t measurementTime_ms = 1500);
     KellerParent(byte modbusAddress, Stream& stream,
              int8_t powerPin, int8_t powerPin2, int8_t enablePin = -1, uint8_t measurementsToAverage = 1,
-             kellerModel model = OTHER, const char *sensName = "Keller-Sensor", int numVariables = 3,
+             kellerModel model = OTHER, const char *sensName = "Keller-Sensor", uint8_t numVariables = 3,
              uint32_t warmUpTime_ms = 500, uint32_t stabilizationTime_ms = 5000, uint32_t measurementTime_ms = 1500);
     virtual ~KellerParent();
 
@@ -63,7 +63,7 @@ private:
     kellerModel _model;
     byte _modbusAddress;
     Stream* _stream;
-    int _RS485EnablePin;
+    int8_t _RS485EnablePin;
     int8_t _powerPin2;
 };
